+++ tests/Functional/ConfigFunctionalTest.php
@@ -47,4 +47,23 @@ class ConfigFunctionalTest extends OdsXlsXlsxFunctionalTest
 
         static::assertDirectoryExists($extension->getAttributes()['cache']['xml'], 'Cache directory does not exist');
     }
+
+    /**
+     * @throws \Exception
+     */
+    public function testCsvWriterAttributes()
+    {
+        /**
+         * @var TwigSpreadsheetExtension $extension
+         */
+        $extension = static::$kernel->getContainer()->get('mewes_k_twig_spreadsheet.twig_spreadsheet_extension');
+
+        static::assertEquals(';', $extension->getAttributes()['csv_writer']['delimiter'], 'Unexpected attribute');
+        static::assertEquals('\'', $extension->getAttributes()['csv_writer']['enclosure'], 'Unexpected attribute');
+        static::assertFalse($extension->getAttributes()['csv_writer']['excel_compatibility'], 'Unexpected attribute');
+        static::assertFalse($extension->getAttributes()['csv_writer']['include_separator_line'], 'Unexpected attribute');
+        static::assertEquals("\r\n", $extension->getAttributes()['csv_writer']['line_ending'], 'Unexpected attribute');
+        static::assertEquals(0, $extension->getAttributes()['csv_writer']['sheet_index'], 'Unexpected attribute');
+        static::assertTrue($extension->getAttributes()['csv_writer']['use_bom'], 'Unexpected attribute');
+    }
 }