+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  * Fixed issue #13 'Incorrect cache directory for images' (technetium)
  * Fixed 'xlsmergestyles' which merged non-array values instead of overriding them
  * Added 'xlscellindex' and 'xlsrowindex' functions to get the current position
+ * Added advanced CSV writer configurations like custom value separators etc
  * Improved internal index handling
  * Improved code performance